Add CI test-site deployment job
This change adds a non-voting job to validate the test-site in
openstack-helm-deployments. Currently, the phases defined in
openstack-helm-deployments point to airshipctl documents to provide a
base infrastructure for testing functions/composites provided by
openstack-helm-deployments.

Once a workload is added to openstack-helm-deployments, the workload
script can be added to this test suite.

Depends-On: https://review.opendev.org/754729

Change-Id: Ie9f4056a98147320059bc8e98fdcf87b250797c7