Add memcached function and OSH composite
This change adds a memcached function that deploys the OpenStack-Helm
memcached chart [0] and adds the function to a new composite named
OpenStack-Helm. The OpenStack-Helm composite will serve as a releasable collection
of documents that enables operators to deploy OpenStack-Helm using
Kustomize [1] and the Helm Operator [2]. Additionally, the
OpenStack-Helm composite can be used as a Kustomize entrypoint for
Airship, enabling operators to deploy OpenStack-Helm using airshipctl
[3].

This change adds the OpenStack-Helm composite to the target cluster
workload phase in order to validate the composite using the existing
airshipctl testing infrastructure.

[0] https://opendev.org/openstack/openstack-helm-infra
[1] https://github.com/kubernetes-sigs/kustomize
[2] https://github.com/fluxcd/helm-operator
[3] https://opendev.org/airship/airshipctl

Change-Id: I571f8c85028199f4fce3613a03b4812d97906bc9
Signed-off-by: Drew Walters <andrew.walters@att.com>